+++ lib/elastic_apm/context/request.rb
@@ -4,8 +4,6 @@ module ElasticAPM
   class Context
     # @api private
     class Request
-      include NaivelyHashable
-
       attr_accessor :body, :cookies, :env, :headers, :http_version, :method,
         :socket, :url
     end