+++ lib/elastic_apm/context/request/url.rb
@@ -7,8 +7,6 @@ module ElasticAPM
     class Request
       # @api private
       class Url
-        include NaivelyHashable
-
         SKIPPED_PORTS = {
           'http' => 80,
           'https' => 443