+++ lib/elastic_apm/metadata/process_info.rb
@@ -1,26 +1,18 @@
 # frozen_string_literal: true
 
 module ElasticAPM
-  module Metadata
+  class Metadata
     # @api private
     class ProcessInfo
       def initialize(config)
         @config = config
+
+        @argv = ARGV
+        @pid = $PID || Process.pid
+        @title = $PROGRAM_NAME
       end
 
-      def build
-        pid = $PID || Process.pid
-        return unless pid
-        {
-          argv: ARGV,
-          pid: pid,
-          title: $PROGRAM_NAME
-        }
-      end
-
-      def self.build(config)
-        new(config).build
-      end
+      attr_reader :argv, :pid, :title
     end
   end
 end