+++ lib/elastic_apm/transport/serializers.rb
@@ -15,10 +15,22 @@ module ElasticAPM
           @config = config
         end
 
+        attr_reader :config
+
         private
 
         def ms(micros)
           micros.to_f / 1_000
+        end
+
+        def keyword_field(value)
+          Util.truncate(value)
+        end
+
+        def keyword_object(hash)
+          hash.tap do |h|
+            h.each { |k, v| hash[k] = keyword_field(v) }
+          end
         end
       end
 
@@ -28,10 +40,12 @@ module ElasticAPM
           @transaction = Serializers::TransactionSerializer.new(config)
           @span = Serializers::SpanSerializer.new(config)
           @error = Serializers::ErrorSerializer.new(config)
+          @metadata = Serializers::MetadataSerializer.new(config)
         end
 
-        attr_reader :transaction, :span, :error
+        attr_reader :transaction, :span, :error, :metadata
 
+        # rubocop:disable Metrics/MethodLength
         def serialize(resource)
           case resource
           when Transaction
@@ -40,10 +54,13 @@ module ElasticAPM
             span.build(resource)
           when Error
             error.build(resource)
+          when Metadata
+            metadata.build(resource)
           else
             raise UnrecognizedResource, resource.inspect
           end
         end
+        # rubocop:enable Metrics/MethodLength
       end
 
       def self.new(config)
@@ -53,6 +70,8 @@ module ElasticAPM
   end
 end
 
-require 'elastic_apm/transport/serializers/transaction_serializer'
-require 'elastic_apm/transport/serializers/span_serializer'
+require 'elastic_apm/transport/serializers/context_serializer'
 require 'elastic_apm/transport/serializers/error_serializer'
+require 'elastic_apm/transport/serializers/metadata_serializer'
+require 'elastic_apm/transport/serializers/span_serializer'
+require 'elastic_apm/transport/serializers/transaction_serializer'