+++ lib/elastic_apm/transport/serializers/error_serializer.rb
@@ -5,6 +5,10 @@ module ElasticAPM
     module Serializers
       # @api private
       class ErrorSerializer < Serializer
+        def context_serializer
+          @context_serializer ||= ContextSerializer.new(config)
+        end
+
         # rubocop:disable Metrics/MethodLength
         def build(error)
           base = {
@@ -15,7 +19,7 @@ module ElasticAPM
 
             culprit: error.culprit,
             timestamp: error.timestamp,
-            context: error.context.to_h
+            context: context_serializer.build(error.context)
           }
 
           if (exception = error.exception)
@@ -35,9 +39,9 @@ module ElasticAPM
         def build_exception(exception)
           {
             message: exception.message,
-            type: exception.type,
-            module: exception.module,
-            code: exception.code,
+            type: keyword_field(exception.type),
+            module: keyword_field(exception.module),
+            code: keyword_field(exception.code),
             attributes: exception.attributes,
             stacktrace: exception.stacktrace.to_a,
             handled: exception.handled
@@ -47,9 +51,9 @@ module ElasticAPM
         def build_log(log)
           {
             message: log.message,
-            level: log.level,
-            logger_name: log.logger_name,
-            param_message: log.param_message,
+            level: keyword_field(log.level),
+            logger_name: keyword_field(log.logger_name),
+            param_message: keyword_field(log.param_message),
             stacktrace: log.stacktrace.to_a
           }
         end