+++ spec/elastic_apm/context/user_spec.rb
@@ -4,20 +4,20 @@ module ElasticAPM
   RSpec.describe Context::User do
     it 'sets values from passed object' do
       PossiblyUser = Struct.new(:id, :email, :username)
-      record = PossiblyUser.new(1, 'a@a', 'monkeyface')
+      record = PossiblyUser.new(1, 'a@a', 'abe')
 
       user = described_class.new(Config.new, record)
-      expect(user.to_h).to eq(
-        id: 1,
-        email: 'a@a',
-        username: 'monkeyface'
-      )
+      expect(user.id).to eq '1'
+      expect(user.email).to eq 'a@a'
+      expect(user.username).to eq 'abe'
     end
 
     it "doesn't explode with missing methods" do
       expect do
         user = described_class.new(Config.new, Object.new)
-        expect(user.to_h).to eq(id: nil, email: nil, username: nil)
+        expect(user.id).to be_nil
+        expect(user.email).to be_nil
+        expect(user.username).to be_nil
       end.to_not raise_exception
     end
   end