+++ spec/elastic_apm/metadata/system_info_spec.rb
@@ -1,23 +1,19 @@
 # frozen_string_literal: true
 
 module ElasticAPM
-  module Metadata
-    RSpec.describe SystemInfo do
-      describe '#build' do
-        subject { described_class.new(Config.new).build }
+  RSpec.describe Metadata::SystemInfo do
+    describe '#initialize' do
+      subject { described_class.new(Config.new) }
 
-        it { should be_a Hash }
-
-        it 'has values' do
-          %i[hostname architecture platform].each do |key|
-            expect(subject[key]).to_not be_nil
-          end
+      it 'has values' do
+        %i[hostname architecture platform].each do |key|
+          expect(subject.send(key)).to_not be_nil
         end
+      end
 
-        context 'hostname' do
-          it 'has no newline at the end' do
-            expect(subject[:hostname]).not_to match(/\n\z/)
-          end
+      context 'hostname' do
+        it 'has no newline at the end' do
+          expect(subject.hostname).not_to match(/\n\z/)
         end
       end
     end